+++ change_log.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2024-03-26
+
+## Added
+
+* Tests are now displayed graphically in the debug information
+* Page history is now shown as test cases, can easily copy them to clipboard
+* Improved error messages for incorrect types in page content
+* Buttons can now take Arguments, and Arguments can be embedded as hidden inputs.
+
+
 ## [Unreleased]
 
 ## [0.2.0] - 2023-09-22